[AI] Move print_utils to common_utils
